main/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
mysqlclient==2.2.7
django-mysql==4.14.0
django-cors-headers==4.3.1
pika==1.1.0
djangorestframework-simplejwt==5.5.1
django-extensions==3.2.3